#18 add coverage and pytest. Rename files in more pythonic way
