+++ requirements-dev.txt
@@ -1,4 +1,5 @@
 astroid==4.0.4
+coverage==7.13.4
 dill==0.4.1
 isort==7.0.0
 mccabe==0.7.0
