+++ src/exception/invalid_path_exception.py
